+++ addon/controllers/lookup-dialog.js
@@ -113,6 +113,14 @@ export default ListFormController.extend(SortableRouteMixin, {
   */
   reloadObserverIsActive: false,
 
+  /**
+    Service that triggers lookup events.
+
+    @property lookupEventsService
+    @type Service
+  */
+  lookupEventsService: Ember.inject.service('lookup-events'),
+
   actions: {
     /**
       Handlers OLV row click, Save selected row to object master property and close modal window.
@@ -133,6 +141,7 @@ export default ListFormController.extend(SortableRouteMixin, {
     */
     createdModalDialog(modalDialog) {
       this.set('_openedModalDialog', modalDialog);
+      this.get('lookupEventsService').lookupDialogOnVisibleTrigger(this.get('componentName'), modalDialog);
     },
 
     /**
@@ -177,7 +186,8 @@ export default ListFormController.extend(SortableRouteMixin, {
       sizeClass: this.get('sizeClass'),
       saveTo: this.get('saveTo'),
       currentLookupRow: this.get('currentLookupRow'),
-      customPropertiesData: this.get('customPropertiesData')
+      customPropertiesData: this.get('customPropertiesData'),
+      componentName: this.get('componentName')
     };
 
     reloadDataHandler(this.get('reloadContext'), reloadData);
@@ -222,7 +232,7 @@ export default ListFormController.extend(SortableRouteMixin, {
     @private
   */
   _selectMaster(master) {
-    var saveTo = this.get('saveTo');
+    let saveTo = this.get('saveTo');
     if (!saveTo) {
       throw new Error('Don\'t know where to save - no saveTo data defined.');
     }
@@ -245,5 +255,7 @@ export default ListFormController.extend(SortableRouteMixin, {
       openedDialog.modal('hide');
       this.set('_openedModalDialog', undefined);
     }
+
+    this.get('lookupEventsService').lookupDialogOnHiddenTrigger(this.get('componentName'));
   },
 });