+++ CHANGELOG.md
@@ -3,9 +3,26 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* `flexberry-lookup` component:
+    * Add support for change related model.
+    * Add `lookup-events` service with triggers: lookupDialogOnShowTrigger, lookupDialogOnVisibleTrigger, lookupDialogOnHiddenTrigger.
+    * Now the user can not press the choose button several times before the modal dialog is open. NOTE: Need to add componenName.
+
 ### Fixed
-* `flexberry-simpleolv` component: 
+* `flexberry-simpleolv` component:
+    * Double rendering.
+    * Now data loading indicator and placeholder are never shows at the same time.
     * Fix `resetFilters` action.
+
+### Removed
+* `flexberry-simpleolv` component:
+    * Remove `_attributeChanged` observers.
+
+## [0.8.0-beta.3] - 2016-10-31
+### Changed
+* Replace `a` tag with `link-to` helper in `sitemap-node` template, because `a` tag with `href-to` heper causes full page reload.
+* Updated dependency on `ember-flexberry-data` addon to v0.7.1-beta.2 version.
 
 ## [0.8.0-beta.2] - 2016-10-28
 ### Added